+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # MapTiler AR Control Changelog
 
+## 2.1.0
+### New Features
+- Now working also with Capacitor inside a mobile app (iOS only)
+### Others
+- Optimization of the UMD bundle size
+
+
 ## 2.0.1
 ### Bug Fixes
 - Fixed: when displayed bound were wider than 180° the Haversine distance used was from the back of Earth, resulting in a too narrow 3D model